Ensure we escape sequences of different entities
In a naive, multi-pass entity replacement (eg, replace all `&` with
`&amp;`, replace all `<` with `&lt;`) the replacement order is
important. (You must replace `&` with `&amp;` first, lest you replace
`<` with `&lt;` then replace `&lt;` with `&amp;lt;`.)

The `escapeHTML` function is a single-pass replacement of each entity at
a time, so is not vulnerable to such a failure mode, but add a test to
avoid regressions.